Switched place recognition & bindings and Calonder descriptor bindings to use byte signatures.
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@9396 f5854215-dd47-0410-b2c4-cdd35faa7885

+++ vision/calonder_descriptor/src/py.cpp
@@ -13,8 +13,8 @@
 
 using namespace features;
 
-typedef float SigType;
-//typedef uint8_t SigType;
+//typedef float SigType;
+typedef uint8_t SigType;
 typedef Promote<SigType>::type DistanceType;
 
 typedef struct {
